Move CodeMirrorRainlang component in ui-components
