+++ tauri-app/src/routes/orders/add/+page.svelte
@@ -23,8 +23,7 @@
   } from '$lib/services/order';
   import { ethersExecute } from '$lib/services/ethersTx';
   import { formatEthersTransactionError } from '$lib/utils/transaction';
-  import CodeMirrorRainlang from '$lib/components/CodeMirrorRainlang.svelte';
-  import { promiseTimeout } from '@rainlanguage/ui-components';
+  import { promiseTimeout, CodeMirrorRainlang } from '@rainlanguage/ui-components';
   import { SentrySeverityLevel, reportErrorToSentry } from '$lib/services/sentry';
   import { pickScenarios } from '$lib/services/pickConfig';
   import {
@@ -274,7 +273,11 @@
       >
         {#each Array.from($generatedRainlang.entries()) as [scenario, rainlangText]}
           <TabItem bind:open={openTab[scenario.name]} title={scenario.name}>
-            <CodeMirrorRainlang bind:value={rainlangText} disabled={true} />
+            <CodeMirrorRainlang
+              bind:value={rainlangText}
+              disabled={true}
+              codeMirrorTheme={$codeMirrorTheme}
+            />
           </TabItem>
         {/each}
       </Tabs>
